+++ src/api/authApi.ts
@@ -1,0 +1,54 @@
+import { BaseApi } from "./baseApi";
+
+export interface SigninData {
+  login: string;
+  password: string;
+}
+
+export interface SignupData {
+  first_name: string;
+  second_name: string;
+  login: string;
+  email: string;
+  password: string;
+  phone: string;
+}
+
+export interface User {
+  id: number;
+  first_name: string;
+  second_name: string;
+  login: string;
+  email: string;
+  password: string;
+  phone: string;
+  avatar: string;
+}
+
+class AuthApi extends BaseApi {
+  constructor() {
+    super("/auth");
+  }
+
+  signin(data: SigninData) {
+    return this.http.post("/signin", data);
+  }
+
+  signup(data: SignupData) {
+    return this.http.post("/signup", data);
+  }
+
+  read(): Promise<User> {
+    return this.http.get("/user");
+  }
+
+  logout() {
+    return this.http.post("/logout");
+  }
+
+  create = undefined;
+  update = undefined;
+  delete = undefined;
+}
+
+export { AuthApi };
